Standardise Crate Startup Functions
