+++ crates/refunder/src/main.rs
@@ -1,14 +1,4 @@
-use clap::Parser;
-
 #[tokio::main]
 async fn main() {
-    let args = refunder::arguments::Arguments::parse();
-    observe::tracing::initialize(
-        args.logging.log_filter.as_str(),
-        args.logging.log_stderr_threshold,
-    );
-    observe::panic_hook::install();
-    tracing::info!("running refunder with validated arguments:\n{}", args);
-    observe::metrics::setup_registry(Some("refunder".into()), None);
-    refunder::main(args).await;
+    refunder::start(std::env::args()).await
 }
