+++ crates/solvers/src/main.rs
@@ -1,7 +1,4 @@
 #[tokio::main]
 async fn main() {
-    observe::panic_hook::install();
-
-    // TODO implement Display for the arguments
-    solvers::run::run(std::env::args(), None).await;
+    solvers::start(std::env::args()).await;
 }
